fix(ci): harden workflow text validation
